Dara QueryServices now direct sublasses of QueryService; fix for query creation
