+++ src/main/java/io/github/infolis/resolve/QueryService.java
@@ -9,6 +9,8 @@ import io.github.infolis.model.BaseModel;
 import io.github.infolis.model.entity.Entity;
 import io.github.infolis.model.entity.SearchResult;
 
+import java.net.MalformedURLException;
+import java.net.URL;
 import java.util.List;
 import java.util.Set;
 
@@ -19,8 +21,8 @@ import java.util.Set;
  */
 @JsonTypeInfo(use = Id.CLASS,include = JsonTypeInfo.As.PROPERTY,property = "type")
 @JsonSubTypes({
-    @Type(value = HTMLQueryService.class),
-    @Type(value = SolrQueryService.class),
+    @Type(value = DaraHTMLQueryService.class),
+    @Type(value = DaraSolrQueryService.class),
     })
 public abstract class QueryService extends BaseModel {
 
@@ -30,7 +32,8 @@ public abstract class QueryService extends BaseModel {
     protected String target = "";
     private double reliability =0.0;
     protected Set<QueryField> queryStrategy;
-
+    protected int maxNumber = 1000;
+    
     public QueryService(String target) {
         this.target = target;
     }
@@ -39,6 +42,8 @@ public abstract class QueryService extends BaseModel {
         this.target = target;
         this.reliability = reliability;
     }
+    
+    public abstract URL createQuery(Entity entity) throws MalformedURLException;
 
     public abstract List<SearchResult> find(Entity entity);
 
@@ -61,6 +66,20 @@ public abstract class QueryService extends BaseModel {
      */
     public void setReliability(double reliability) {
         this.reliability = reliability;
+    }
+    
+    /**
+     * @return the maxNumber
+     */
+    public int getMaxNumber() {
+        return maxNumber;
+    }
+
+    /**
+     * @param maxNumber the maxNumber to set
+     */
+    public void setMaxNumber(int maxNumber) {
+        this.maxNumber = maxNumber;
     }
     
     public void setQueryStrategy(Set<QueryField> queryStrategy) {
